+++ CHANGELOG.md
@@ -5,6 +5,10 @@ Here is the list of the changes and features provided by the Rainbow-CLI
 
 All notable changes to Rainbow-CLI will be documented in this file.
 
+### [1.44.0] - 2018-07-08
+- RQRAINB-1008 Create application on behalf and change ownership of an application
+- RQRAINB-1009: Renew the application secret
+
 ### [1.43.1] - 2018-07-08
 - Switch to Rainbow-node-sdk 1.43.2 to fix Node6 issue